+++ src/Illuminate/Foundation/Providers/FoundationServiceProvider.php
@@ -2,16 +2,81 @@
 
 namespace Illuminate\Foundation\Providers;
 
-use Illuminate\Support\AggregateServiceProvider;
+use Illuminate\Routing\Redirector;
+use Illuminate\Support\ServiceProvider;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Routing\Events\RouteMatched;
+use Symfony\Component\HttpFoundation\Request;
+use Illuminate\Contracts\Validation\ValidatesWhenResolved;
 
-class FoundationServiceProvider extends AggregateServiceProvider
+class FoundationServiceProvider extends ServiceProvider
 {
     /**
-     * The provider class names.
+     * Register the service provider.
      *
-     * @var array
+     * @return void
      */
-    protected $providers = [
-        FormRequestServiceProvider::class,
-    ];
+    public function register()
+    {
+        //
+    }
+
+    /**
+     * Initialize the form request with data from the given request.
+     *
+     * @param  \Illuminate\Foundation\Http\FormRequest  $form
+     * @param  \Symfony\Component\HttpFoundation\Request  $current
+     * @return void
+     */
+    public function boot()
+    {
+        $this->configureFormRequests();
+    }
+
+    /**
+     * Configure the form request related services.
+     *
+     * @return void
+     */
+    protected function configureFormRequests()
+    {
+        $this->app->afterResolving(function (ValidatesWhenResolved $resolved) {
+            $resolved->validate();
+        });
+
+        $this->app['events']->listen(RouteMatched::class, function () {
+            $this->app->resolving(function (FormRequest $request, $app) {
+                $this->initializeRequest($request, $app['request']);
+
+                $request->setContainer($app)->setRedirector($app->make(Redirector::class));
+            });
+        });
+    }
+
+    /**
+     * Initialize the form request with data from the given request.
+     *
+     * @param  \Illuminate\Foundation\Http\FormRequest  $form
+     * @param  \Symfony\Component\HttpFoundation\Request  $current
+     * @return void
+     */
+    protected function initializeRequest(FormRequest $form, Request $current)
+    {
+        $files = $current->files->all();
+
+        $files = is_array($files) ? array_filter($files) : $files;
+
+        $form->initialize(
+            $current->query->all(), $current->request->all(), $current->attributes->all(),
+            $current->cookies->all(), $files, $current->server->all(), $current->getContent()
+        );
+
+        if ($session = $current->getSession()) {
+            $form->setSession($session);
+        }
+
+        $form->setUserResolver($current->getUserResolver());
+
+        $form->setRouteResolver($current->getRouteResolver());
+    }
 }
