+++ src/Illuminate/Validation/ValidationServiceProvider.php
@@ -3,10 +3,16 @@
 namespace Illuminate\Validation;
 
 use Illuminate\Support\ServiceProvider;
-use Illuminate\Contracts\Validation\ValidatesWhenResolved;
 
 class ValidationServiceProvider extends ServiceProvider
 {
+    /**
+     * Indicates if loading of the provider is deferred.
+     *
+     * @var bool
+     */
+    protected $defer = true;
+
     /**
      * Register the service provider.
      *
@@ -14,23 +20,9 @@ class ValidationServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        $this->registerValidationResolverHook();
-
         $this->registerPresenceVerifier();
 
         $this->registerValidationFactory();
-    }
-
-    /**
-     * Register the "ValidatesWhenResolved" container hook.
-     *
-     * @return void
-     */
-    protected function registerValidationResolverHook()
-    {
-        $this->app->afterResolving(function (ValidatesWhenResolved $resolved) {
-            $resolved->validate();
-        });
     }
 
     /**
@@ -64,5 +56,17 @@ class ValidationServiceProvider extends ServiceProvider
         $this->app->singleton('validation.presence', function ($app) {
             return new DatabasePresenceVerifier($app['db']);
         });
+    }
+
+    /**
+     * Get the services provided by the provider.
+     *
+     * @return array
+     */
+    public function provides()
+    {
+        return [
+            'validator', 'validation.presence',
+        ];
     }
 }
